Fix loading conditional for waiting props; Move Loading to utils (#8)
